requirements.txt
streamlit
tensorflow==2.19.0
numpy
pillow
pandas
